[ refactor ] : app.css 삭제 및 globalstyle폴더변경
