--- src/App.css
@@ -1,11 +1,0 @@
-#root {
-  max-width: max-content;
-  width: 100%;
-  height: 100%;
-  margin: 0 auto;
-  height: 100%;
-  width: 100%;
-  text-align: center;
-}
-
-
