+++ src/style/GlobalStyle.jsx
@@ -9,6 +9,15 @@ const GlobalStyle = createGlobalStyle`
   
   color: pink;
   
+  #root {
+  max-width: max-content;
+  width: 100%;
+  height: 100%;
+  margin: 0 auto;
+  height: 100%;
+  width: 100%;
+  text-align: center;
+}
 `;
 
 export default GlobalStyle;
